Remove build of GraphOptimizerParameters from composer_builder.py
